+++ cashflow/templates/kitchen/assembly.html
@@ -1,0 +1,9 @@
+{% extends 'base/base_nav_footer.html' %}
+
+{% block content %}
+  <div class="container-fluid">
+    <div>
+      <h1 class="text-xs-center">Ensamblado!</h1>
+    </div>
+  </div>
+{% endblock %}
