+++ kitchen/admin.py
@@ -1,0 +1,14 @@
+from django.contrib import admin
+
+from kitchen.models import ProcessedCartridge, Warehouse
+
+
+@admin.register(ProcessedCartridge)
+class AdminProcessedCartridge(admin.ModelAdmin):
+    list_display = ('status', 'cartridge_parent', 'package_cartridge_parent', 'created_at')
+
+
+@admin.register(Warehouse)
+class AdminWarehouse(admin.ModelAdmin):
+    list_display = ('supply', 'status', 'quantity', 'waste', 'cost')
+
